% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phecodeX_map}
\alias{phecodeX_map}
\title{phecodeX_map
Dataframe containing phecode mappings for phecodeX}
\format{
\subsection{\code{phecodeX_map}}{

A data frame with 79583 observations on the following 3 variables.
\describe{
\item{\code{vocabulary_id}}{Character vector representing the vocabulary of
the code- always 'ICD10'}
\item{\code{code}}{Character vector representing the specific code to be
mapped}
\item{\code{phecode}}{Character vector representing the direct phecode
mapping}
...
}
}
}
\usage{
phecodeX_map
}
\description{
This data frame maps each ICD9 and ICD10 code to its directly mapped phecode(s).
Can be provided to \code{\link[PheWAS:createPhenotypes]{createPhenotypes}}
in the \code{vocabulary.map} parameter to convert codes for use with phecodeX in
the PheWAS methods.
}
\keyword{datasets}
