% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phecodeX_rollup_map}
\alias{phecodeX_rollup_map}
\title{PhecodeX Rollup Map}
\format{
\subsection{\code{phecodeX_rollup_map}}{

a data frame with 3319 rows of 2 columns
\describe{
\item{code}{the reference phecode}
\item{phecode_unrolled}{the parent phecode}
...
}
}
}
\usage{
phecodeX_rollup_map
}
\description{
a data map that contains phecode X parent-child pairings
}
\keyword{datasets}
