% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phecodeX_sex}
\alias{phecodeX_sex}
\title{PheinfoX sex exclusions map}
\format{
\subsection{\code{phecodeX_sex}}{

a data frame with 3612 rows of 3 columns
\describe{
\item{phecode}{the reference phecode}
\item{male_only}{a boolean which is TRUE for male-specific phecodes}
\item{female_only}{a boolean which is TRUE for male-specific phecodes}
...
}
}
}
\usage{
phecodeX_sex
}
\description{
a data map that contains sex exclusion information on each phecode
}
\keyword{datasets}
