% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pheinfo_X}
\alias{pheinfo_X}
\title{PheinfoX map}
\format{
\subsection{\code{pheinfo_X}}{

a data frame with 3612 rows of 6 columns
\describe{
\item{phecode}{the reference phecode}
\item{description}{a string describing the phecode}
\item{groupnum}{TBD}
\item{group}{what type of disease the phecode is}
\item{color}{TBD}
...
}
}
}
\usage{
pheinfo_X
}
\description{
a data map that contains information on each phecode
}
\keyword{datasets}
